01: Correct segfault in segfault.asm
I made a typo which caused a segfault. That was a lot sooner than I
expected.

NASM did suggest there was a problem with the file:

```
segfault.asm:11: warning: label alone on a line without a colon might be in error [-w+orphan-labels]
```

So this is saying: I don't know what `sycall` is (it isn't an
instruction). I'm interpreting it as a label, so do with that what you
want.

This is going to be fun :)